change nextjs to ssg
