+++ apps/web/app/_components/theme-providers.tsx
@@ -1,7 +1,8 @@
+/* eslint-disable @nx/enforce-module-boundaries */
 'use client';
 
+import siteMetadata from '@/data/site-metadata';
 import { ThemeProvider } from 'next-themes';
-import siteMetadata from '../../../data/site-metadata';
 
 export function ThemeProviders({ children }: { children: React.ReactNode }) {
   return (
