+++ apps/web/app/blog/tags/[tag]/page.tsx
@@ -1,11 +1,11 @@
+/* eslint-disable @nx/enforce-module-boundaries */
+import siteMetadata from '@/data/site-metadata';
+import ListLayout from '@/layouts/list-layout-with-tags';
 import { allBlogs } from 'contentlayer/generated';
 import { slug } from 'github-slugger';
 import { Metadata } from 'next';
 import { allCoreContent, sortPosts } from 'pliny/utils/contentlayer';
-import siteMetadata from '../../../../../data/site-metadata';
-import { Locale } from '../../../../../i18n/i18n-config';
-import ListLayout from '../../../../../layouts/list-layout-with-tags';
-import tagData from '../../../../../public/tag-data.json';
+import tagData from '../../../../public/tag-data.json';
 import { genPageMetadata } from '../../../_components/seo';
 
 export async function generateMetadata({ params }: { params: { tag: string } }): Promise<Metadata> {
@@ -31,7 +31,7 @@ export const generateStaticParams = async () => {
   return paths;
 };
 
-export default function TagPage({ params }: { params: { tag: string; lang: Locale } }) {
+export default function TagPage({ params }: { params: { tag: string } }) {
   const tag = decodeURI(params.tag);
   // Capitalize first letter and convert space to dash
   const title = tag[0].toUpperCase() + tag.split(' ').join('-').slice(1);
@@ -39,5 +39,5 @@ export default function TagPage({ params }: { params: { tag: string; lang: Local
   const filteredPosts = allCoreContent(
     sortedPosts.filter((post) => post.tags && post.tags.map((t) => slug(t)).includes(tag))
   );
-  return <ListLayout posts={filteredPosts} title={title} locale={params.lang} />;
+  return <ListLayout posts={filteredPosts} title={title} />;
 }
