+++ apps/web/app/sites/page.tsx
@@ -1,5 +1,6 @@
-import Card from '../../..//components/card';
-import sitesData from '../../../data/sites-data';
+/* eslint-disable @nx/enforce-module-boundaries */
+import Card from '@/components/card';
+import sitesData from '@/data/sites-data';
 import { genPageMetadata } from '../_components/seo';
 
 export const metadata = genPageMetadata({ title: 'Games' });
