+++ apps/web/components/comments.tsx
@@ -1,8 +1,9 @@
+/* eslint-disable @nx/enforce-module-boundaries */
 'use client';
 
+import siteMetadata from '@/data/site-metadata';
 import { Comments as CommentsComponent } from 'pliny/comments';
 import { useState } from 'react';
-import siteMetadata from '../data/site-metadata';
 
 export default function Comments({ slug }: { slug: string }) {
   const [loadComments, setLoadComments] = useState(false);
