+++ apps/web/data/header-nav-links.ts
@@ -1,0 +1,10 @@
+const headerNavLinks = [
+  { href: '/', title: 'Home' },
+  { href: '/blog', title: 'Blog' },
+  { href: '/blog/tags', title: 'Tags' },
+  { href: '/sites', title: 'Sites' },
+  { href: '/games', title: 'Games' },
+  { href: '/about', title: 'About' }
+];
+
+export default headerNavLinks;
