--- apps/web/i18n/dictionaries/ja.json
@@ -1,63 +1,0 @@
-{
-  "description": "システム開発とプログラミングに関する情報サイトです。",
-  "header_nav_links": [
-    { "href": "/", "title": "ホーム" },
-    { "href": "/blog", "title": "ブログ" },
-    { "href": "/blog/tags", "title": "タグ" },
-    { "href": "/sites", "title": "サイト" },
-    { "href": "/games", "title": "ゲーム" },
-    { "href": "/about", "title": "About" }
-  ],
-  "navigation": {
-    "menu": "メニュー",
-    "cart": "カート",
-    "shop": "ショップ",
-    "language": "言語"
-  },
-  "product_categories": {
-    "all": "全て",
-    "coats_and_jackets": "コートとジャケット",
-    "dresses": "ドレス",
-    "new_arrivals": "新着アイテム",
-    "pants_and_skirts": "パンツとスカート"
-  },
-  "product": {
-    "size_guide": "サイズガイド",
-    "size": "サイズ",
-    "color": "カラー",
-    "length": "着丈",
-    "add_to_cart": "カートに追加",
-    "buy_now": "今すぐ購入",
-    "note": "注意：各アイテムは丁寧に手作業で製作され、製造に最大5営業日かかることがあります。"
-  },
-  "size_guide": {
-    "close": "閉じる",
-    "size_and_fit": {
-      "title": "サイズとフィット",
-      "fit": "通常のサイズに合います。通常のサイズを選んでください。",
-      "materials": "中重量、伸縮性のない生地",
-      "model_size": "モデルの身長は180 cmで、サイズSおよび「トール」の着丈を着用しています。"
-    },
-    "length": {
-      "title": "着丈",
-      "select_height": "OK KINOの服が完璧にフィットするようにするために、身長を選択するオプションを提供しています。",
-      "petit": "もし身長が164 cm以下であれば、「プティ」を選択してください。",
-      "regular": "もし身長が165-175 cmの間であれば、「レギュラー」を選択してください。",
-      "tall": "もし身長が176 cm以上であれば、「トール」を選択してください。"
-    },
-    "size_guide": {
-      "title": "サイズガイド",
-      "size": "サイズ",
-      "bust": "バスト",
-      "waist": "ウエスト",
-      "hips": "ヒップ"
-    },
-    "how_to_measure": {
-      "title": "測り方：",
-      "guide": "正しいサイズを選ぶために、次のように体を測定してください。",
-      "bust": "バスト — 最も大きな部分を測定",
-      "waist": "ウエスト — 自然なウエストラインを測定",
-      "hips": "ヒップ — 自然なウエストラインから20cm下を測定"
-    }
-  }
-}
