--- apps/web/i18n/get-dictionary.ts
@@ -1,13 +1,0 @@
-import 'server-only';
-import type { Locale } from './i18n-config';
-import { i18n } from './i18n-config';
-
-// We enumerate all dictionaries here for better linting and typescript support
-// We also get the defalt import for cleaner types
-const dictionaries = {
-  en: () => import('./dictionaries/en.json').then((module) => module.default),
-  ja: () => import('./dictionaries/ja.json').then((module) => module.default)
-};
-
-export const getDictionary = async (locale: Locale) =>
-  dictionaries[i18n.locales.includes(locale) ? locale : i18n.defaultLocale]();
