--- apps/web/i18n/i18n-config.ts
@@ -1,6 +1,0 @@
-export const i18n = {
-  defaultLocale: 'en',
-  locales: ['en', 'ja']
-} as const;
-
-export type Locale = (typeof i18n)['locales'][number];
