+++ apps/web/next.config.js
@@ -1,9 +1,6 @@
 const { composePlugins, withNx } = require('@nx/next');
 const { withContentlayer } = require('next-contentlayer');
 
-const dns = require('node:dns');
-dns.setDefaultResultOrder('ipv4first');
-
 /**
  * @type {import('@nx/next/plugins/with-nx').WithNxOptions}
  **/
@@ -14,6 +11,7 @@ const nextConfig = {
     svgr: false
   },
   images: {
+    unoptimized: true,
     domains: ['localhost'],
     remotePatterns: [
       {
@@ -21,8 +19,8 @@ const nextConfig = {
         hostname: String(process.env.ATMAN_WEB_STORAGE_HOSTNAME)
       }
     ]
-  }
-  // swcMinify: true
+  },
+  output: 'export'
 };
 
 const plugins = [
