+++ package.json
@@ -34,6 +34,7 @@
     "@nestjs/jwt": "^10.1.1",
     "@nestjs/passport": "^10.0.1",
     "@nestjs/platform-express": "^10.0.2",
+    "@nestjs/serve-static": "^4.0.0",
     "@prisma/client": "^5.2.0",
     "@swc/helpers": "~0.5.0",
     "@tailwindcss/forms": "^0.5.6",
